+++ CHANGES.md
@@ -2,6 +2,7 @@
 * add setting in `Extra Settings` to hide `Report a Bug`
 * add filter `wpbt_hide_report_a_bug`
 * update if `Report a Bug` plugin active
+* link to settings if `Report a Bug` is hidden
 
 #### 3.5.0 / 2023-06-30
 * update dashboard widget for MarComm publishing of posts